requirements/base.txt
uvicorn[standard]==0.18.3
gunicorn==20.1.0
Django==3.2
python-dotenv==0.20.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
django-filter==22.1
Pillow==9.2.0
redis==4.3.4
# twisted-iocpsupport==1.0.1

# news
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf == 0.22.5
sentry-sdk==1.14.0
python-logstash==0.4.8

selenium==4.8.2
webdriver_manager==3.8.5
bs4==0.0.1
celery==5.2.7
django-better-admin-arrayfield==1.4.2 
ultimate-sitemap-parser==0.5
lxml
django-baton==2.6.2
django-import-export==3.1.0
django-celery-beat==2.5.0
django-pgtrigger==4.6.0
django-redis
django-modeltranslation